@@ -596,7 +596,9 @@ class _ChannelListViewState extends State<ChannelListView> {
               child: ChannelPreview(
                 onLongPress: widget.onChannelLongPress,
                 channel: channel,
-                onImageTap: () => widget.onImageTap?.call(channel),
+                onImageTap: widget.onImageTap != null
+                    ? () => widget.onImageTap!(channel)
+                    : null,
                 onTap: (channel) => onTap(channel, widget.channelWidget),
               ),
             ),
@@ -609,7 +611,7 @@ class _ChannelListViewState extends State<ChannelListView> {
     if (widget.onChannelTap != null) {
       onTap = widget.onChannelTap!;
     } else {
-      onTap = (client, _) {
+      onTap = (channel, _) {
         if (widget.channelWidget == null) {
           return;
         }
@@ -617,7 +619,7 @@ class _ChannelListViewState extends State<ChannelListView> {
           context,
           MaterialPageRoute(
             builder: (context) => StreamChannel(
-              channel: client,
+              channel: channel,
               child: widget.channelWidget!,
             ),
           ),
