@@ -709,20 +709,21 @@ class _MessageListViewState extends State<MessageListView> {
             );
           },
         ),
-        BetterStreamBuilder<bool>(
-          stream: streamChannel!.channel.state!.isUpToDateStream,
-          initialData: streamChannel!.channel.state!.isUpToDate,
-          builder: (context, snapshot) => ValueListenableBuilder<bool>(
-            valueListenable: _showScrollToBottom,
-            child: _buildScrollToBottom(),
-            builder: (context, value, child) {
-              if (!snapshot || value) {
-                return child!;
-              }
-              return const Offstage();
-            },
+        if (widget.showScrollToBottom)
+          BetterStreamBuilder<bool>(
+            stream: streamChannel!.channel.state!.isUpToDateStream,
+            initialData: streamChannel!.channel.state!.isUpToDate,
+            builder: (context, snapshot) => ValueListenableBuilder<bool>(
+              valueListenable: _showScrollToBottom,
+              child: _buildScrollToBottom(),
+              builder: (context, value, child) {
+                if (!snapshot || value) {
+                  return child!;
+                }
+                return const Offstage();
+              },
+            ),
           ),
-        ),
         if (widget.showFloatingDateDivider)
           _buildFloatingDateDivider(itemCount),
       ],
